Fix wrong calculate function of DPOIndicator and fix DPOIndicatorTest
Because of the addition (index+timeShift) for the ema value the calculate
function threw always a NullpointerException for the last tick.
Regarding the definition (see link in class) the timeShift must
be subtract from ema index.

Additional i have updated the corresponding test class. For some reasons
that i do not understand, the last @test function has been expected a
NullpointerException for an index in range
(timeSeries.getEndIndex()-timeFrame)