Update Indicator.java
Add convenience "static" method "toDouble()" which returns a list of "Doubles" from a corresponding Indicator (based on decimal). With this, it is easy to use 3rd-party libs which only expect the data as double values (for example, commons-math or other mathematical libraries needs a double[] array as a input argument).